requirements.txt
torch
vaderSentiment
langdetect
feedparser
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
statsmodels
tensorflow
PyYAML
requests
streamlit
plotly
websocket-client
certifi
yfinance
scipy

